# nodejs-edu/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: 'Test'
  description: 'Test'
host: localhost:8080
basePath: /api
paths: 
...
  /cities/{id}:
    delete:
      tags:
       - Cities
      description: Delete selected city from db
      summary: delete
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: deleted
    put: 
      tags:
        - Cities
      description: Create or update city
      summary: create or update
      consumes: 
        - application/x-www-form-urlencoded
      parameters:
        - name: id
          in: path
          required: true
          type: string
          default: 5a3acd9e909333809453c73d
        - name: name
          in: formData
          type: string
          default: Minsk
          required: true
        - name: country
          in: formData
          type: string
          default: Belarus
          required: true
        - name: capital
          in: formData
          type: string
          default: "false"
          required: true
        - name: location.lat
          in: formData
          type: string
          default: "10"
          required: true
        - name: location.lon
          in: formData
          type: string
          default: "11"
          required: true
      responses:
        200:
          description: -updated or created.
...